+++ src/Phorkus.Consensus/CommonCoin/CommonCoin.cs
@@ -88,7 +88,7 @@ namespace Phorkus.Consensus.CommonCoin
                 Validator = new Validator
                 {
                     // TODO: somehow fill validator field
-                    ValidatorIndex = (ulong) GetMyId(),
+                    ValidatorIndex = GetMyId(),
                     Era = _coinId.Era
                 },
                 Coin = new CommonCoinMessage
